Now the bluetable isn't a fxml file, the Worktable.java is now the class that create and handle all code below workTable.
Create a basic contextMenu, using the ContextMenu.java class, it's working and using the menu.css file how styleClass, but the menu aren't created in current spot.

Created a basic Wire tool, in Wire.java class, now we are able to create basics wire, the tool and the wire is currently in the same file, but the tool will be moved to another file.

Basic OOP class and interfaces creates, created a Component.java to represent any electrical component, and a container to put and register all components creates in some screen.