Changes about how Grid ,Components, and GridPin interact itself.
Modify the Component.java to accept multi GridPin.java, now the GridPin is array, and is abstracted to give the first element.

Add BiPinComponent.java, is the same that Component.java, but the array can be accessed to get the first or second pin.

Modify the cable to work with new Components class;

Bug to fix:
->Source don't show in current spot.